+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Respect the custom IntelliJ Platform configuration name when creating a custom task using the local IntelliJ Platform path
 - Properly share the requested IntelliJ Platform between internal Gradle plugin components
 - Refactor `CollectorTransformer` to exclude tests-related jars using `filterNot` as `minus` doesn't remove duplicates
+- Refactor `JarSearchableOptionsTask` to improve the `searchableOption` outgoing artifact registration JetBrains/intellij-platform-gradle-plugin#1968
 
 ## [2.6.0] - 2025-05-14